Included Additional DLL Dependencies. GL / GLFW Libraries Added for Window rendering and OpenGL SFML Window Additionally added. Updated RenderWindow to Window
- Previously RenderWindow is early 2012 version which is not supported, and it has now been upgraded for modern opengl.

Currently Only Renders a white Box in the middle of the Screen, which I'm not sure if it is supposed to render anything further or not.

The Project has also been compiled to include ISO C++17 Standard (/std:c++17) Instead of Newer versions, to use the project.

IT is important to change it before Debugging!!
Or you will mess up the project and have to re-pull the repository or go through fixing it in notepad.

The project Expects that Windows 10 SDKs Are downloaded within your computer as well as Graphics Card update is made otherwise it would not compile!

Get SDKs from Visual studio Setup
Additional -> windows 10 SDK.